feature(Notifications): Documenting NotificationModel
