--- DidIFloss/Service/Notification/Notification.swift
@@ -1,17 +1,0 @@
-//
-//  NotificationManager.swift
-//  DidIFloss
-//
-//  Created by Lucas Migge on 08/01/24.
-//
-
-import Foundation
-import UserNotifications
-
-
-///  Namespace for gruping  features related to user notification
-///
-/// Here you should find the service for schedule notifications and model for diferent types of Notifications
-struct Notification {
-    
-}
